Fix clinia icons stories
