+++ docs/stories/clinia-icons.stories.tsx
@@ -1,750 +1,9 @@
-import { AccessibilityIcon } from "@clinia-ui/icons/AccessibilityIcon";
-import { AddIcon } from "@clinia-ui/icons/AddIcon";
-import { AddImageIcon } from "@clinia-ui/icons/AddImageIcon";
-import { AddSuggestionIcon } from "@clinia-ui/icons/AddSuggestionIcon";
-import { AdminIcon } from "@clinia-ui/icons/AdminIcon";
-import { AgeFocusIcon } from "@clinia-ui/icons/AgeFocusIcon";
-import { AnalyticsIcon } from "@clinia-ui/icons/AnalyticsIcon";
-import { AnalyzeModuleIcon } from "@clinia-ui/icons/AnalyzeModuleIcon";
-import { AreaDownIcon } from "@clinia-ui/icons/AreaDownIcon";
-import { AreaIcon } from "@clinia-ui/icons/AreaIcon";
-import { AreaLeftIcon } from "@clinia-ui/icons/AreaLeftIcon";
-import { AreaRightIcon } from "@clinia-ui/icons/AreaRightIcon";
-import { ArrowDownIcon } from "@clinia-ui/icons/ArrowDownIcon";
-import { ArrowLeftCircleIcon } from "@clinia-ui/icons/ArrowLeftCircleIcon";
-import { ArrowRightCircleIcon } from "@clinia-ui/icons/ArrowRightCircleIcon";
-import { ArrowUpIcon } from "@clinia-ui/icons/ArrowUpIcon";
-import { BookingIcon } from "@clinia-ui/icons/BookingIcon";
-import { BookmarkIcon } from "@clinia-ui/icons/BookmarkIcon";
-import { BrokenImageIcon } from "@clinia-ui/icons/BrokenImageIcon";
-import { BulletPointsIcon } from "@clinia-ui/icons/BulletPointsIcon";
-import { ByNameIcon } from "@clinia-ui/icons/ByNameIcon";
-import { ByTypeIcon } from "@clinia-ui/icons/ByTypeIcon";
-import { CalendarIcon } from "@clinia-ui/icons/CalendarIcon";
-import { CaseToReferIcon } from "@clinia-ui/icons/CaseToReferIcon";
-import { CheckboxIcon } from "@clinia-ui/icons/CheckboxIcon";
-import { CheckmarkIcon } from "@clinia-ui/icons/CheckmarkIcon";
-import { ClearIcon } from "@clinia-ui/icons/ClearIcon";
 import { CliniaIcon } from "@clinia-ui/icons/CliniaIcon";
-import { ClinicIcon } from "@clinia-ui/icons/ClinicIcon";
-import { ClockIcon } from "@clinia-ui/icons/ClockIcon";
-import { CloseIcon } from "@clinia-ui/icons/CloseIcon";
-import { CopyIcon } from "@clinia-ui/icons/CopyIcon";
-import { CorrespondanceIcon } from "@clinia-ui/icons/CorrespondanceIcon";
-import { CustomFieldIcon } from "@clinia-ui/icons/CustomFieldIcon";
-import { CustomFieldsDisabledIcon } from "@clinia-ui/icons/CustomFieldsDisabledIcon";
-import { CustomFieldsEnabledIcon } from "@clinia-ui/icons/CustomFieldsEnabledIcon";
-import { DataIcon } from "@clinia-ui/icons/DataIcon";
-import { DateTimeIcon } from "@clinia-ui/icons/DateTimeIcon";
-import { DeactivateIcon } from "@clinia-ui/icons/DeactivateIcon";
-import { DegreeIcon } from "@clinia-ui/icons/DegreeIcon";
-import { DeleteIcon } from "@clinia-ui/icons/DeleteIcon";
-import { DepartmentIcon } from "@clinia-ui/icons/DepartmentIcon";
-import { DeselectIcon } from "@clinia-ui/icons/DeselectIcon";
-import { DetailsIcon } from "@clinia-ui/icons/DetailsIcon";
-import { DevConsoleIcon } from "@clinia-ui/icons/DevConsoleIcon";
-import { DirectionsIcon } from "@clinia-ui/icons/DirectionsIcon";
-import { DirectoryIcon } from "@clinia-ui/icons/DirectoryIcon";
-import { DoctorIcon } from "@clinia-ui/icons/DoctorIcon";
-import { DoubleLeftArrowIcon } from "@clinia-ui/icons/DoubleLeftArrowIcon";
-import { DoubleRightArrowIcon } from "@clinia-ui/icons/DoubleRightArrowIcon";
-import { DownloadIcon } from "@clinia-ui/icons/DownloadIcon";
-import { DragIcon } from "@clinia-ui/icons/DragIcon";
-import { DriveThroughIcon } from "@clinia-ui/icons/DriveThroughIcon";
-import { DropdownIcon } from "@clinia-ui/icons/DropdownIcon";
-import { EditIcon } from "@clinia-ui/icons/EditIcon";
-import { ElderlyIcon } from "@clinia-ui/icons/ElderlyIcon";
-import { EllipseIcon } from "@clinia-ui/icons/EllipseIcon";
-import { EmailIcon } from "@clinia-ui/icons/EmailIcon";
-import { ExportIcon } from "@clinia-ui/icons/ExportIcon";
-import { ExternalLinkIcon } from "@clinia-ui/icons/ExternalLinkIcon";
-import { FaceIcon } from "@clinia-ui/icons/FaceIcon";
-import { FaxIcon } from "@clinia-ui/icons/FaxIcon";
-import { FeaturedIcon } from "@clinia-ui/icons/FeaturedIcon";
-import { FeedbackIcon } from "@clinia-ui/icons/FeedbackIcon";
-import { FilterActiveIcon } from "@clinia-ui/icons/FilterActiveIcon";
-import { FilterInactiveIcon } from "@clinia-ui/icons/FilterInactiveIcon";
-import { FiltersIcon } from "@clinia-ui/icons/FiltersIcon";
-import { FlagActiveIcon } from "@clinia-ui/icons/FlagActiveIcon";
-import { FlagInactiveIcon } from "@clinia-ui/icons/FlagInactiveIcon";
-import { ForbiddenIcon } from "@clinia-ui/icons/ForbiddenIcon";
-import { GenderIcon } from "@clinia-ui/icons/GenderIcon";
-import { GraduationIcon } from "@clinia-ui/icons/GraduationIcon";
-import { GraphIcon } from "@clinia-ui/icons/GraphIcon";
-import { HealthResourceIcon } from "@clinia-ui/icons/HealthResourceIcon";
-import { HelpIcon } from "@clinia-ui/icons/HelpIcon";
-import { HomeIcon } from "@clinia-ui/icons/HomeIcon";
-import { HospitalIcon } from "@clinia-ui/icons/HospitalIcon";
-import { IdentifiantIcon } from "@clinia-ui/icons/IdentifiantIcon";
-import { InfoIcon } from "@clinia-ui/icons/InfoIcon";
-import { InformationIcon } from "@clinia-ui/icons/InformationIcon";
-import { InsuranceIcon } from "@clinia-ui/icons/InsuranceIcon";
-import { IntegrationIcon } from "@clinia-ui/icons/IntegrationIcon";
-import { KebabMenuIcon } from "@clinia-ui/icons/KebabMenuIcon";
-import { LanguagesIcon } from "@clinia-ui/icons/LanguagesIcon";
-import { LeadIcon } from "@clinia-ui/icons/LeadIcon";
-import { LinkIcon } from "@clinia-ui/icons/LinkIcon";
-import { ListNbIcon } from "@clinia-ui/icons/ListNbIcon";
-import { ListViewIcon } from "@clinia-ui/icons/ListViewIcon";
-import { LiveIcon } from "@clinia-ui/icons/LiveIcon";
-import { LocationIcon } from "@clinia-ui/icons/LocationIcon";
-import { LogoutIcon } from "@clinia-ui/icons/LogoutIcon";
-import { LongTextIcon } from "@clinia-ui/icons/LongTextIcon";
-import { ManageUsersIcon } from "@clinia-ui/icons/ManageUsersIcon";
-import { MapIcon } from "@clinia-ui/icons/MapIcon";
-import { MeatballMenuIcon } from "@clinia-ui/icons/MeatballMenuIcon";
-import { MedicalTeamIcon } from "@clinia-ui/icons/MedicalTeamIcon";
-import { MembersIcon } from "@clinia-ui/icons/MembersIcon";
-import { MembersInvitedIcon } from "@clinia-ui/icons/MembersInvitedIcon";
-import { ModalityIcon } from "@clinia-ui/icons/ModalityIcon";
-import { NewPatientsIcon } from "@clinia-ui/icons/NewPatientsIcon";
-import { NoteIcon } from "@clinia-ui/icons/NoteIcon";
-import { NotificationIcon } from "@clinia-ui/icons/NotificationIcon";
-import { NoViewIcon } from "@clinia-ui/icons/NoViewIcon";
-import { NumberIcon } from "@clinia-ui/icons/NumberIcon";
-import { NumeroIcon } from "@clinia-ui/icons/NumeroIcon";
-import { OpenInNewTabIcon } from "@clinia-ui/icons/OpenInNewTabIcon";
-import { OrganizationIcon } from "@clinia-ui/icons/OrganizationIcon";
-import { PagerIcon } from "@clinia-ui/icons/PagerIcon";
-import { PDFIcon } from "@clinia-ui/icons/PDFIcon";
-import { PeopleIcon } from "@clinia-ui/icons/PeopleIcon";
-import { PhoneIcon } from "@clinia-ui/icons/PhoneIcon";
-import { PracticePlaceIcon } from "@clinia-ui/icons/PracticePlaceIcon";
-import { PreviewIcon } from "@clinia-ui/icons/PreviewIcon";
-import { ProductPanelIcon } from "@clinia-ui/icons/ProductPanelIcon";
-import { ProductWorkspaceIcon } from "@clinia-ui/icons/ProductWorkspaceIcon";
-import { ProfessionIcon } from "@clinia-ui/icons/ProfessionIcon";
-import { ProfileIcon } from "@clinia-ui/icons/ProfileIcon";
-import { ProviderIcon } from "@clinia-ui/icons/ProviderIcon";
-import { RefreshIcon } from "@clinia-ui/icons/RefreshIcon";
-import { ResendInviteIcon } from "@clinia-ui/icons/ResendInviteIcon";
-import { SchoolIcon } from "@clinia-ui/icons/SchoolIcon";
-import { SearchEmptyStateIcon } from "@clinia-ui/icons/SearchEmptyStateIcon";
-import { SearchIcon } from "@clinia-ui/icons/SearchIcon";
-import { SecurityIcon } from "@clinia-ui/icons/SecurityIcon";
-import { ServiceOfferedIcon } from "@clinia-ui/icons/ServiceOfferedIcon";
-import { ServicesIcon } from "@clinia-ui/icons/ServicesIcon";
-import { SettingsIcon } from "@clinia-ui/icons/SettingsIcon";
-import { ShareIcon } from "@clinia-ui/icons/ShareIcon";
-import { ShortTextIcon } from "@clinia-ui/icons/ShortTextIcon";
-import { ShortZAIcon } from "@clinia-ui/icons/ShortZAIcon";
-import { SortAZIcon } from "@clinia-ui/icons/SortAZIcon";
-import { StarRatingIcon } from "@clinia-ui/icons/StarRatingIcon";
-import { SublevelIcon } from "@clinia-ui/icons/SublevelIcon";
-import { SubscriptionsIcon } from "@clinia-ui/icons/SubscriptionsIcon";
-import { SuggestionIcon } from "@clinia-ui/icons/SuggestionIcon";
-import { SuggestionRemovalIcon } from "@clinia-ui/icons/SuggestionRemovalIcon";
-import { TileViewIcon } from "@clinia-ui/icons/TileViewIcon";
-import { ToddlerFaceIcon } from "@clinia-ui/icons/ToddlerFaceIcon";
-import { TooltipIcon } from "@clinia-ui/icons/TooltipIcon";
-import { TrendDownIcon } from "@clinia-ui/icons/TrendDownIcon";
-import { TrendUpIcon } from "@clinia-ui/icons/TrendUpIcon";
-import { UploadIcon } from "@clinia-ui/icons/UploadIcon";
-import { VideosIcon } from "@clinia-ui/icons/VideosIcon";
-import { ViewIcon } from "@clinia-ui/icons/ViewIcon";
-import { WaitingTimeIcon } from "@clinia-ui/icons/WaitingTimeIcon";
-import { WalkinIcon } from "@clinia-ui/icons/WalkinIcon";
-import { WebsiteIcon } from "@clinia-ui/icons/WebsiteIcon";
-import {
-  Button,
-  Card,
-  CardBody,
-  Center,
-  Stack,
-  Text,
-  Tooltip,
-} from "@clinia-ui/react";
-import { Meta } from "@storybook/react";
-import { useState } from "react";
-
-const a: Meta = {
-  decorators: [(story) => <>{story.toString()}</>],
-};
-
-type IconStory = () => [string, JSX.Element];
-
-const IconDecorator = ({ story }: { story: IconStory }) => {
-  const [name, component] = story();
-  const [open, setOpen] = useState(false);
-
-  return (
-    <Tooltip isOpen={open} label={`'${name}' copied to clipboard`}>
-      <Button
-        height="fit-content"
-        variant="unstyled"
-        onClick={() => {
-          navigator.clipboard.writeText(name);
-          setOpen(true);
-          setTimeout(() => setOpen(false), 2000);
-        }}
-      >
-        <Card align="center" width={170}>
-          <CardBody>
-            <Stack>
-              <Center>{component}</Center>
-              <Text fontSize="xs">{name}</Text>
-            </Stack>
-          </CardBody>
-        </Card>
-      </Button>
-    </Tooltip>
-  );
-};
 
 export default {
   title: "Icons",
-  decorators: [(story: IconStory) => <IconDecorator story={story} />],
 };
 
-const props = {
-  color: "primary.main",
-  boxSize: "2em",
-};
-
-export const AccessibilityIcon_ = () => [
-  "AccessibilityIcon",
-  <AccessibilityIcon {...props} />,
-];
-
-export const AddIcon_ = () => ["AddIcon", <AddIcon {...props} />];
-
-export const AddImageIcon_ = () => [
-  "AddImageIcon",
-  <AddImageIcon {...props} />,
-];
-
-export const AddSuggestionIcon_ = () => [
-  "AddSuggestionIcon",
-  <AddSuggestionIcon {...props} />,
-];
-
-export const AdminIcon_ = () => ["AdminIcon", <AdminIcon {...props} />];
-export const AgeFocusIcon_ = () => [
-  "AgeFocusIcon",
-  <AgeFocusIcon {...props} />,
-];
-
-export const AnalyticsIcon_ = () => [
-  "AnalyticsIcon",
-  <AnalyticsIcon {...props} />,
-];
-
-export const AnalyzeModuleIcon_ = () => [
-  "AnalyzeModuleIcon",
-  <AnalyzeModuleIcon {...props} />,
-];
-
-export const AreaDownIcon_ = () => [
-  "AreaDownIcon",
-  <AreaDownIcon {...props} />,
-];
-
-export const AreaIcon_ = () => ["AreaIcon", <AreaIcon {...props} />];
-
-export const AreaLeftIcon_ = () => [
-  "AreaLeftIcon",
-  <AreaLeftIcon {...props} />,
-];
-
-export const AreaRightIcon_ = () => [
-  "AreaRightIcon",
-  <AreaRightIcon {...props} />,
-];
-
-export const ArrowDownIcon_ = () => [
-  "ArrowDownIcon",
-  <ArrowDownIcon {...props} />,
-];
-
-export const ArrowLeftCircleIcon_ = () => [
-  "ArrowLeftCircleIcon",
-  <ArrowLeftCircleIcon {...props} />,
-];
-
-export const ArrowRightCircleIcon_ = () => [
-  "ArrowRightCircleIcon",
-  <ArrowRightCircleIcon {...props} />,
-];
-
-export const ArrowUpIcon_ = () => ["ArrowUpIcon", <ArrowUpIcon {...props} />];
-
-export const BookingIcon_ = () => ["BookingIcon", <BookingIcon {...props} />];
-
-export const BookmarkIcon_ = () => [
-  "BookmarkIcon",
-  <BookmarkIcon {...props} />,
-];
-
-export const BrokenImageIcon_ = () => [
-  "BrokenImageIcon",
-  <BrokenImageIcon {...props} />,
-];
-
-export const BulletPointsIcon_ = () => [
-  "BulletPointsIcon",
-  <BulletPointsIcon {...props} />,
-];
-
-export const ByNameIcon_ = () => ["ByNameIcon", <ByNameIcon {...props} />];
-
-export const ByTypeIcon_ = () => ["ByTypeIcon", <ByTypeIcon {...props} />];
-
-export const CalendarIcon_ = () => [
-  "CalendarIcon",
-  <CalendarIcon {...props} />,
-];
-
-export const CaseToReferIcon_ = () => [
-  "CaseToReferIcon",
-  <CaseToReferIcon {...props} />,
-];
-
-export const CheckboxIcon_ = () => [
-  "CheckboxIcon",
-  <CheckboxIcon {...props} />,
-];
-
-export const CheckmarkIcon_ = () => [
-  "CheckmarkIcon",
-  <CheckmarkIcon {...props} />,
-];
-
-export const ClearIcon_ = () => ["ClearIcon", <ClearIcon {...props} />];
-
-export const CliniaIcon_ = () => ["CliniaIcon", <CliniaIcon {...props} />];
-
-export const ClinicIcon_ = () => ["ClinicIcon", <ClinicIcon {...props} />];
-
-export const ClockIcon_ = () => ["ClockIcon", <ClockIcon {...props} />];
-
-export const CloseIcon_ = () => ["CloseIcon", <CloseIcon {...props} />];
-
-export const CopyIcon_ = () => ["CopyIcon", <CopyIcon {...props} />];
-
-export const CorrespondanceIcon_ = () => [
-  "CorrespondanceIcon",
-  <CorrespondanceIcon {...props} />,
-];
-
-export const CustomFieldIcon_ = () => [
-  "CustomFieldIcon",
-  <CustomFieldIcon {...props} />,
-];
-
-export const CustomFieldsDisabledIcon_ = () => [
-  "CustomFieldsDisabledIcon",
-  <CustomFieldsDisabledIcon {...props} />,
-];
-
-export const CustomFieldsEnabledIcon_ = () => [
-  "CustomFieldsEnabledIcon",
-  <CustomFieldsEnabledIcon {...props} />,
-];
-
-export const DataIcon_ = () => ["DataIcon", <DataIcon {...props} />];
-
-export const DateTimeIcon_ = () => [
-  "DateTimeIcon",
-  <DateTimeIcon {...props} />,
-];
-
-export const DeactivateIcon_ = () => [
-  "DeactivateIcon",
-  <DeactivateIcon {...props} />,
-];
-
-export const DegreeIcon_ = () => ["DegreeIcon", <DegreeIcon {...props} />];
-
-export const DeleteIcon_ = () => ["DeleteIcon", <DeleteIcon {...props} />];
-
-export const DepartmentIcon_ = () => [
-  "DepartmentIcon",
-  <DepartmentIcon {...props} />,
-];
-
-export const DeselectIcon_ = () => [
-  "DeselectIcon",
-  <DeselectIcon {...props} />,
-];
-
-export const DetailsIcon_ = () => ["DetailsIcon", <DetailsIcon {...props} />];
-
-export const DevConsoleIcon_ = () => [
-  "DevConsoleIcon",
-  <DevConsoleIcon {...props} />,
-];
-
-export const DirectionsIcon_ = () => [
-  "DirectionsIcon",
-  <DirectionsIcon {...props} />,
-];
-
-export const DirectoryIcon_ = () => [
-  "DirectoryIcon",
-  <DirectoryIcon {...props} />,
-];
-
-export const DoctorIcon_ = () => ["DoctorIcon", <DoctorIcon {...props} />];
-
-export const DoubleLeftArrowIcon_ = () => [
-  "DoubleLeftArrowIcon",
-  <DoubleLeftArrowIcon {...props} />,
-];
-
-export const DoubleRightArrowIcon_ = () => [
-  "DoubleRightArrowIcon",
-  <DoubleRightArrowIcon {...props} />,
-];
-
-export const DownloadIcon_ = () => [
-  "DownloadIcon",
-  <DownloadIcon {...props} />,
-];
-
-export const DragIcon_ = () => ["DragIcon", <DragIcon {...props} />];
-
-export const DriveThroughIcon_ = () => [
-  "DriveThroughIcon",
-  <DriveThroughIcon {...props} />,
-];
-
-export const DropdownIcon_ = () => [
-  "DropdownIcon",
-  <DropdownIcon {...props} />,
-];
-
-export const EditIcon_ = () => ["EditIcon", <EditIcon {...props} />];
-
-export const ElderlyIcon_ = () => ["ElderlyIcon", <ElderlyIcon {...props} />];
-
-export const EllipseIcon_ = () => ["EllipseIcon", <EllipseIcon {...props} />];
-
-export const EmailIcon_ = () => ["EmailIcon", <EmailIcon {...props} />];
-
-export const ExportIcon_ = () => ["ExportIcon", <ExportIcon {...props} />];
-
-export const ExternalLinkIcon_ = () => [
-  "ExternalLinkIcon",
-  <ExternalLinkIcon {...props} />,
-];
-
-export const FaceIcon_ = () => ["FaceIcon", <FaceIcon {...props} />];
-
-export const FaxIcon_ = () => ["FaxIcon", <FaxIcon {...props} />];
-
-export const FeaturedIcon_ = () => [
-  "FeaturedIcon",
-  <FeaturedIcon {...props} />,
-];
-
-export const FeedbackIcon_ = () => [
-  "FeedbackIcon",
-  <FeedbackIcon {...props} />,
-];
-
-export const FilterActiveIcon_ = () => [
-  "FilterActiveIcon",
-  <FilterActiveIcon {...props} />,
-];
-
-export const FilterInactiveIcon_ = () => [
-  "FilterInactiveIcon",
-  <FilterInactiveIcon {...props} />,
-];
-
-export const FiltersIcon_ = () => ["FiltersIcon", <FiltersIcon {...props} />];
-
-export const FlagActiveIcon_ = () => [
-  "FlagActiveIcon",
-  <FlagActiveIcon {...props} />,
-];
-
-export const FlagInactiveIcon_ = () => [
-  "FlagInactiveIcon",
-  <FlagInactiveIcon {...props} />,
-];
-
-export const ForbiddenIcon_ = () => [
-  "ForbiddenIcon",
-  <ForbiddenIcon {...props} />,
-];
-
-export const GenderIcon_ = () => ["GenderIcon", <GenderIcon {...props} />];
-
-export const GraduationIcon_ = () => [
-  "GraduationIcon",
-  <GraduationIcon {...props} />,
-];
-
-export const GraphIcon_ = () => ["GraphIcon", <GraphIcon {...props} />];
-
-export const HealthResourceIcon_ = () => [
-  "HealthResourceIcon",
-  <HealthResourceIcon {...props} />,
-];
-
-export const HelpIcon_ = () => ["HelpIcon", <HelpIcon {...props} />];
-
-export const HomeIcon_ = () => ["HomeIcon", <HomeIcon {...props} />];
-
-export const HospitalIcon_ = () => [
-  "HospitalIcon",
-  <HospitalIcon {...props} />,
-];
-
-export const IdentifiantIcon_ = () => [
-  "IdentifiantIcon",
-  <IdentifiantIcon {...props} />,
-];
-
-export const InfoIcon_ = () => ["InfoIcon", <InfoIcon {...props} />];
-
-export const InformationIcon_ = () => [
-  "InformationIcon",
-  <InformationIcon {...props} />,
-];
-
-export const InsuranceIcon_ = () => [
-  "InsuranceIcon",
-  <InsuranceIcon {...props} />,
-];
-
-export const IntegrationIcon_ = () => [
-  "IntegrationIcon",
-  <IntegrationIcon {...props} />,
-];
-
-export const KebabMenuIcon_ = () => [
-  "KebabMenuIcon",
-  <KebabMenuIcon {...props} />,
-];
-
-export const LanguagesIcon_ = () => [
-  "LanguagesIcon",
-  <LanguagesIcon {...props} />,
-];
-
-export const LeadIcon_ = () => ["LeadIcon", <LeadIcon {...props} />];
-
-export const LinkIcon_ = () => ["LinkIcon", <LinkIcon {...props} />];
-
-export const ListNbIcon_ = () => ["ListNbIcon", <ListNbIcon {...props} />];
-
-export const ListViewIcon_ = () => [
-  "ListViewIcon",
-  <ListViewIcon {...props} />,
-];
-
-export const LiveIcon_ = () => ["LiveIcon", <LiveIcon {...props} />];
-
-export const LocationIcon_ = () => [
-  "LocationIcon",
-  <LocationIcon {...props} />,
-];
-
-export const LogoutIcon_ = () => ["LogoutIcon", <LogoutIcon {...props} />];
-
-export const LongTextIcon_ = () => [
-  "LongTextIcon",
-  <LongTextIcon {...props} />,
-];
-
-export const ManageUsersIcon_ = () => [
-  "ManageUsersIcon",
-  <ManageUsersIcon {...props} />,
-];
-
-export const MapIcon_ = () => ["MapIcon", <MapIcon {...props} />];
-
-export const MeatballMenuIcon_ = () => [
-  "MeatballMenuIcon",
-  <MeatballMenuIcon {...props} />,
-];
-
-export const MedicalTeamIcon_ = () => [
-  "MedicalTeamIcon",
-  <MedicalTeamIcon {...props} />,
-];
-
-export const MembersIcon_ = () => ["MembersIcon", <MembersIcon {...props} />];
-
-export const MembersInvitedIcon_ = () => [
-  "MembersInvitedIcon",
-  <MembersInvitedIcon {...props} />,
-];
-
-export const ModalityIcon_ = () => [
-  "ModalityIcon",
-  <ModalityIcon {...props} />,
-];
-
-export const NewPatientsIcon_ = () => [
-  "NewPatientsIcon",
-  <NewPatientsIcon {...props} />,
-];
-
-export const NoteIcon_ = () => ["NoteIcon", <NoteIcon {...props} />];
-
-export const NotificationIcon_ = () => [
-  "NotificationIcon",
-  <NotificationIcon {...props} />,
-];
-
-export const NoViewIcon_ = () => ["NoViewIcon", <NoViewIcon {...props} />];
-
-export const NumberIcon_ = () => ["NumberIcon", <NumberIcon {...props} />];
-
-export const NumeroIcon_ = () => ["NumeroIcon", <NumeroIcon {...props} />];
-
-export const OpenInNewTabIcon_ = () => [
-  "OpenInNewTabIcon",
-  <OpenInNewTabIcon {...props} />,
-];
-
-export const OrganizationIcon_ = () => [
-  "OrganizationIcon",
-  <OrganizationIcon {...props} />,
-];
-
-export const PagerIcon_ = () => ["PagerIcon", <PagerIcon {...props} />];
-
-export const PDFIcon_ = () => ["PDFIcon", <PDFIcon {...props} />];
-
-export const PeopleIcon_ = () => ["PeopleIcon", <PeopleIcon {...props} />];
-
-export const PhoneIcon_ = () => ["PhoneIcon", <PhoneIcon {...props} />];
-
-export const PracticePlaceIcon_ = () => [
-  "PracticePlaceIcon",
-  <PracticePlaceIcon {...props} />,
-];
-
-export const PreviewIcon_ = () => ["PreviewIcon", <PreviewIcon {...props} />];
-
-export const ProductPanelIcon_ = () => [
-  "ProductPanelIcon",
-  <ProductPanelIcon {...props} />,
-];
-
-export const ProductWorkspaceIcon_ = () => [
-  "ProductWorkspaceIcon",
-  <ProductWorkspaceIcon {...props} />,
-];
-
-export const ProfessionIcon_ = () => [
-  "ProfessionIcon",
-  <ProfessionIcon {...props} />,
-];
-
-export const ProfileIcon_ = () => ["ProfileIcon", <ProfileIcon {...props} />];
-
-export const ProviderIcon_ = () => [
-  "ProviderIcon",
-  <ProviderIcon {...props} />,
-];
-
-export const RefreshIcon_ = () => ["RefreshIcon", <RefreshIcon {...props} />];
-
-export const ResendInviteIcon_ = () => [
-  "ResendInviteIcon",
-  <ResendInviteIcon {...props} />,
-];
-
-export const SchoolIcon_ = () => ["SchoolIcon", <SchoolIcon {...props} />];
-
-export const SearchEmptyStateIcon_ = () => [
-  "SearchEmptyStateIcon",
-  <SearchEmptyStateIcon {...props} />,
-];
-
-export const SearchIcon_ = () => ["SearchIcon", <SearchIcon {...props} />];
-
-export const SecurityIcon_ = () => [
-  "SecurityIcon",
-  <SecurityIcon {...props} />,
-];
-
-export const ServiceOfferedIcon_ = () => [
-  "ServiceOfferedIcon",
-  <ServiceOfferedIcon {...props} />,
-];
-
-export const ServicesIcon_ = () => [
-  "ServicesIcon",
-  <ServicesIcon {...props} />,
-];
-
-export const SettingsIcon_ = () => [
-  "SettingsIcon",
-  <SettingsIcon {...props} />,
-];
-
-export const ShareIcon_ = () => ["ShareIcon", <ShareIcon {...props} />];
-
-export const ShortTextIcon_ = () => [
-  "ShortTextIcon",
-  <ShortTextIcon {...props} />,
-];
-
-export const ShortZAIcon_ = () => ["ShortZAIcon", <ShortZAIcon {...props} />];
-
-export const SortAZIcon_ = () => ["SortAZIcon", <SortAZIcon {...props} />];
-
-export const StarRatingIcon_ = () => [
-  "StarRatingIcon",
-  <StarRatingIcon {...props} />,
-];
-
-export const SublevelIcon_ = () => [
-  "SublevelIcon",
-  <SublevelIcon {...props} />,
-];
-
-export const SubscriptionsIcon_ = () => [
-  "SubscriptionsIcon",
-  <SubscriptionsIcon {...props} />,
-];
-
-export const SuggestionIcon_ = () => [
-  "SuggestionIcon",
-  <SuggestionIcon {...props} />,
-];
-
-export const SuggestionRemovalIcon_ = () => [
-  "SuggestionRemovalIcon",
-  <SuggestionRemovalIcon {...props} />,
-];
-
-export const TileViewIcon_ = () => [
-  "TileViewIcon",
-  <TileViewIcon {...props} />,
-];
-
-export const ToddlerFaceIcon_ = () => [
-  "ToddlerFaceIcon",
-  <ToddlerFaceIcon {...props} />,
-];
-
-export const TooltipIcon_ = () => ["TooltipIcon", <TooltipIcon {...props} />];
-
-export const TrendDownIcon_ = () => [
-  "TrendDownIcon",
-  <TrendDownIcon {...props} />,
-];
-
-export const TrendUpIcon_ = () => ["TrendUpIcon", <TrendUpIcon {...props} />];
-
-export const UploadIcon_ = () => ["UploadIcon", <UploadIcon {...props} />];
-
-export const VideosIcon_ = () => ["VideosIcon", <VideosIcon {...props} />];
-
-export const ViewIcon_ = () => ["ViewIcon", <ViewIcon {...props} />];
-
-export const WaitingTimeIcon_ = () => [
-  "WaitingTimeIcon",
-  <WaitingTimeIcon {...props} />,
-];
-
-export const WalkinIcon_ = () => ["WalkinIcon", <WalkinIcon {...props} />];
-
-export const WebsiteIcon_ = () => ["WebsiteIcon", <WebsiteIcon {...props} />];
+export const ExampleIcon = () => (
+  <CliniaIcon color="primary.main" boxSize="2em" />
+);
